refactor: Consolidate ChartWithInteractiveFilters from preview and published charts
