+++ app/components/chart-preview.tsx
@@ -2,25 +2,21 @@ import { Trans } from "@lingui/macro";
 import { Box, Theme, Typography } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import Head from "next/head";
-import * as React from "react";
 import { useMemo } from "react";
 
 import { DataSetTable } from "@/browse/datatable";
-import { ChartDataFilters } from "@/charts/shared/chart-data-filters";
-import useSyncInteractiveFilters from "@/charts/shared/use-sync-interactive-filters";
 import { ChartErrorBoundary } from "@/components/chart-error-boundary";
-import { ChartFiltersList } from "@/components/chart-filters-list";
 import { ChartFootnotes } from "@/components/chart-footnotes";
 import {
   ChartTablePreviewProvider,
   useChartTablePreview,
 } from "@/components/chart-table-preview";
-import GenericChart from "@/components/common-chart";
+import { ChartWithInteractiveFilters } from "@/components/common-chart";
 import DebugPanel from "@/components/debug-panel";
 import Flex from "@/components/flex";
 import { HintYellow } from "@/components/hint";
 import { MetadataPanel } from "@/components/metadata-panel";
-import { ChartConfig, DataSource, useConfiguratorState } from "@/configurator";
+import { DataSource, useConfiguratorState } from "@/configurator";
 import {
   useComponentsQuery,
   useDataCubeMetadataQuery,
@@ -29,13 +25,14 @@ import { DataCubePublicationStatus } from "@/graphql/resolver-types";
 import { useLocale } from "@/locales/use-locale";
 import useEvent from "@/utils/use-event";
 
-export const ChartPreview = ({
-  dataSetIri,
-  dataSource,
-}: {
+type ChartPreviewProps = {
   dataSetIri: string;
   dataSource: DataSource;
-}) => {
+};
+
+export const ChartPreview = (props: ChartPreviewProps) => {
+  const { dataSetIri, dataSource } = props;
+
   return (
     <ChartTablePreviewProvider>
       <ChartPreviewInner dataSetIri={dataSetIri} dataSource={dataSource} />
@@ -60,13 +57,8 @@ const useStyles = makeStyles<Theme>({
   },
 });
 
-export const ChartPreviewInner = ({
-  dataSetIri,
-  dataSource,
-}: {
-  dataSetIri: string;
-  dataSource: DataSource;
-}) => {
+export const ChartPreviewInner = (props: ChartPreviewProps) => {
+  const { dataSetIri, dataSource } = props;
   const [state, dispatch] = useConfiguratorState();
   const locale = useLocale();
   const classes = useStyles();
@@ -214,6 +206,7 @@ export const ChartPreviewInner = ({
                   dataSet={dataSetIri}
                   dataSource={dataSource}
                   chartConfig={state.chartConfig}
+                  published={false}
                 />
               )}
             </Box>
@@ -225,41 +218,10 @@ export const ChartPreviewInner = ({
                 onToggleTableView={handleToggleTableView}
               />
             )}
-            <DebugPanel configurator={true} interactiveFilters={true} />
+            <DebugPanel configurator interactiveFilters />
           </>
         )}
       </ChartErrorBoundary>
     </Flex>
   );
 };
-
-type ChartWithInteractiveFiltersProps = {
-  dataSet: string;
-  dataSource: DataSource;
-  chartConfig: ChartConfig;
-};
-
-const ChartWithInteractiveFilters = React.forwardRef(
-  (props: ChartWithInteractiveFiltersProps, ref) => {
-    useSyncInteractiveFilters(props.chartConfig);
-
-    return (
-      <Flex
-        ref={ref}
-        sx={{
-          flexDirection: "column",
-          justifyContent: "space-between",
-          flexGrow: 1,
-        }}
-      >
-        {props.chartConfig.interactiveFiltersConfig?.dataFilters.active ? (
-          <ChartDataFilters {...props} />
-        ) : (
-          <ChartFiltersList {...props} />
-        )}
-        <GenericChart {...props} published={false} />
-      </Flex>
-    );
-  }
-);
-ChartWithInteractiveFilters.displayName = "ChartWithInteractiveFilters";
