+++ app/components/common-chart.tsx
@@ -1,6 +1,11 @@
 import dynamic from "next/dynamic";
+import React from "react";
 
+import { ChartDataFilters } from "@/charts/shared/chart-data-filters";
 import { useQueryFilters } from "@/charts/shared/chart-helpers";
+import useSyncInteractiveFilters from "@/charts/shared/use-sync-interactive-filters";
+import { ChartFiltersList } from "@/components/chart-filters-list";
+import Flex from "@/components/flex";
 import { ChartConfig, DataSource } from "@/configurator";
 
 const ChartAreasVisualization = dynamic(
@@ -101,4 +106,35 @@ const GenericChart = (props: GenericChartProps) => {
   }
 };
 
-export default GenericChart;
+type ChartWithInteractiveFiltersProps = {
+  dataSet: string;
+  dataSource: DataSource;
+  chartConfig: ChartConfig;
+  published: boolean;
+};
+
+export const ChartWithInteractiveFilters = React.forwardRef<
+  HTMLDivElement,
+  ChartWithInteractiveFiltersProps
+>((props, ref) => {
+  useSyncInteractiveFilters(props.chartConfig);
+
+  return (
+    <Flex
+      ref={ref}
+      sx={{
+        flexDirection: "column",
+        justifyContent: "space-between",
+        flexGrow: 1,
+      }}
+    >
+      {props.chartConfig.interactiveFiltersConfig?.dataFilters.active ? (
+        <ChartDataFilters {...props} />
+      ) : (
+        <ChartFiltersList {...props} />
+      )}
+      <GenericChart {...props} />
+    </Flex>
+  );
+});
+ChartWithInteractiveFilters.displayName = "ChartWithInteractiveFilters";
